requirements.txt
langchain==0.0.190
openai==0.27.7
llama-cpp-python==0.1.50
gpt4all==0.2.3
urllib3==2.0.2
SQLAlchemy==2.0.15
chromadb==0.3.23
python-dotenv==1.0.0
pydantic==1.10.8
duckdb==0.8.0
guidance==0.0.61
protobuf==3.20.3
nvitop==1.1.2